requirements.txt
Flask
werkzeug
psycopg2
pdf2image
google-generativeai
KYC_doc
api_maneger